fix: rename props, rename "LinkToSocials" component and related styles
